Update issue 5 Changed default value in GUI as well as in command-line parameters. Fixed problem with cleaning (parameter was ignored).

+++ Osm2garminAPI/src/main/java/org/mantlik/osm2garmin/Region.java
@@ -101,7 +101,7 @@ public class Region {
             mapsdir = mapsdir + "/";
         }
         dir = new File(mapsdir + this.name);
-        if (dir.exists()) {
+        if (dir.exists() && deleteOldMap) {
             Osm2garmin.deleteFile(dir);
         }
         if (!dir.exists()) {
